Extracted subcomponent TimeSeriesLine
